refactor: 글쓰기 페이지에서 필터링하는 Section 컴포넌트 제거(필터 섹션은 컴포넌트로 분리하지 않는 것으로 결정)
1. useCheckbox와 로직 맞추기위해 수행한 이전 커밋은 오히려 코드가 너무 복잡해짐  (CheckboxContainer에서 해주는 onChange 작업을 두번하게되고, 상태가 분산됨)
2. 각 필터 섹션은 CheckboxContainer 자체로 캡슐화 되어 있어서  분리하지 않아도 된다고 판단함
3. writeIdea할 때 선택한 checkboxValue.branches,checkboxValue.purposes에서  Id 배열만 보내기 위해 매핑함
